feat: Initial Commit
Source was created by ChatGPT4o from the second prompt.

First prompt:

What rust crates can be used to get the list of commits for a  crate
in a workspace. I'd like an API that returns comprehensive information
like git log returns but so includes Pull Request information.

Second prompt:

What I'm specifically trying to do is to create a changelog. I'd like
to list all commits associated with a particular subdirectory in a
workspace and identify which Pull Request each commit was associated
with. Which tools could accomplish that?

Third prompt:

Lets try git2 + octocrab lets write  a rust app that is executed from the
command line in the root directory that is a clone of a github repo.
I want it to list in reverse chronological order (newest first) with the
tags in brackets followed by date, "[tag-name] - year-month-day".
After the tag is the list of commits preceded by a dash then the
description, "- description". Where the description is the first line of
the commit message. Finally, if the most recent commit in the history
doesn't have  the preceding tag should be "[unreleased". By default the
output is in markdown format and is written to the console.