wip: Attempt to get commit author
The "correct" approach seems to be to implement "get_commit_with_sha1"
from github and then use code like get_author which returns the
author. But it always does the `fetch_pr_metadata` which is expensive:

 async fn get_author(
    author_map: &mut HashMap<String, String>,
    repo_commit: &RepoCommit,
    octocrab: &Octocrab,
    repo_owner: &str,
    repo_name: &str,
    merge_commit_sha: &str,
 ) -> String {
    let result = fetch_pr_metadata(octocrab, repo_owner, repo_name, merge_commit_sha).await;
    let email = if let Some(commit_author) = repo_commit.commit.author.clone() {
        commit_author.email
    } else {
        "".to_string()
    };

    // I find this obtuse, I wonder what would be clearer?
    let author = author_map.entry(email.clone()).or_insert_with(|| {
        result.map_or("".to_string(), |pr| {
            pr.user.map_or("".to_string(), |user| user.login)
        })
    });

    println!("get_author: author: {author} email: {email}");

    author.clone()
 }

So I think what needs to happen use a commit from git2 to get a commits
sha1 and use that sha1 to get the same commit from github and then use
the author.user.login from this github commit as the "author" in the
changelog. I started doing that with git_commit_with_sha1 and
get_commit_author but it's not working yet.

For the time being I'm going to explore using the github api more directly
rather than octocrab use octocrate where I might be able to fetch
commits from github directly using list_commits with a sha1 page=1 limit=1,
but we'll see.